backend/paper_games/requirements.txt
aioredis==1.3.1
alabaster==0.7.12
asgiref==3.2.7
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.4.2
Automat==20.2.0
Babel==2.8.0
certifi==2020.4.5.1
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
constantly==15.1.0
coverage==5.1
cryptography==2.9.1
daphne==2.5.0
Django==3.0.5
django-cors-headers==3.2.1
django-templated-mail==1.1.1
djangorestframework==3.11.0
djoser==2.0.3
docutils==0.16
hiredis==1.0.1
hyperlink==19.0.0
idna==2.9
imagesize==1.2.0
incremental==17.5.0
Jinja2==2.11.2
MarkupSafe==1.1.1
msgpack==1.0.0
numpy==1.18.3
packaging==20.3
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.6.1
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pyparsing==2.4.7
pytz==2019.3
requests==2.23.0
selenium==3.141.0
service-identity==18.1.0
six==1.14.0
snowballstemmer==2.0.0
Sphinx==3.0.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.9
zope.interface==5.1.0
